db/requirements.txt
pymongo
python-dotenv
pydantic
pydantic-settings
openai